backend/recommender/requirements.txt
Flask
Flask-Cors
pandas
numpy
scikit-learn
python-dotenv
requests